Create follow button component
