+++ app/frontend/app/routes/application.js
@@ -22,24 +22,6 @@ export default Ember.Route.extend({
       }
     },
 
-    toggleFollow: function(user) {
-      if (!this.get('currentUser.isSignedIn')) {
-        alert("Need to be signed in!");
-        return;
-      }
-
-      var originalState = user.get('isFollowed');
-      user.set('isFollowed', !originalState);
-      return ajax({
-        url: "/users/" + user.get('id') + "/follow",
-        type: "POST",
-        dataType: "json"
-      }).then(Ember.K, function() {
-        alert("Something went wrong.");
-        return user.set('isFollowed', originalState);
-      });
-    },
-
     openModal: function(modalName, model) {
       this.controllerFor("modals." + modalName).set('content', model);
       return this.render("modals/" + modalName, {
